Module aws_sdk_fms::error

Modules

internal_error_exception

See InternalErrorException

invalid_input_exception

See InvalidInputException

invalid_operation_exception

See InvalidOperationException

invalid_type_exception

See InvalidTypeException

limit_exceeded_exception

See LimitExceededException

resource_not_found_exception

See ResourceNotFoundException

Structs

AssociateAdminAccountError

DeleteAppsListError

DeleteNotificationChannelError

DeletePolicyError

DeleteProtocolsListError

DisassociateAdminAccountError

GetAdminAccountError

GetAppsListError

GetComplianceDetailError

GetNotificationChannelError

GetPolicyError

GetProtectionStatusError

GetProtocolsListError

GetViolationDetailsError

InternalErrorException

The operation failed because of a system problem, even though the request was valid. Retry

InvalidInputException

The parameters of the request were invalid.

InvalidOperationException

The operation failed because there was nothing to do or the operation wasn't possible. For example, you might have

InvalidTypeException

The value of the Type parameter is invalid.

LimitExceededException

The operation exceeds a resource limit, for example, the maximum number of

ListAppsListsError

ListComplianceStatusError

ListMemberAccountsError

ListPoliciesError

ListProtocolsListsError

ListTagsForResourceError

PutAppsListError

PutNotificationChannelError

PutPolicyError

PutProtocolsListError

ResourceNotFoundException

The specified resource was not found.

TagResourceError

UntagResourceError

Enums

AssociateAdminAccountErrorKind

DeleteAppsListErrorKind

DeleteNotificationChannelErrorKind

DeletePolicyErrorKind

DeleteProtocolsListErrorKind

DisassociateAdminAccountErrorKind

GetAdminAccountErrorKind

GetAppsListErrorKind

GetComplianceDetailErrorKind

GetNotificationChannelErrorKind

GetPolicyErrorKind

GetProtectionStatusErrorKind

GetProtocolsListErrorKind

GetViolationDetailsErrorKind

ListAppsListsErrorKind

ListComplianceStatusErrorKind

ListMemberAccountsErrorKind

ListPoliciesErrorKind

ListProtocolsListsErrorKind

ListTagsForResourceErrorKind

PutAppsListErrorKind

PutNotificationChannelErrorKind

PutPolicyErrorKind

PutProtocolsListErrorKind

TagResourceErrorKind

UntagResourceErrorKind