Module aws_sdk_fms::model

Modules

action_target

See ActionTarget

app

See App

apps_list_data

See AppsListData

apps_list_data_summary

See AppsListDataSummary

aws_ec2_instance_violation

See AwsEc2InstanceViolation

aws_ec2_network_interface_violation

See AwsEc2NetworkInterfaceViolation

aws_vpc_security_group_violation

See AwsVpcSecurityGroupViolation

compliance_violator

See ComplianceViolator

dns_duplicate_rule_group_violation

See DnsDuplicateRuleGroupViolation

dns_rule_group_limit_exceeded_violation

See DnsRuleGroupLimitExceededViolation

dns_rule_group_priority_conflict_violation

See DnsRuleGroupPriorityConflictViolation

ec2_associate_route_table_action

See Ec2AssociateRouteTableAction

ec2_copy_route_table_action

See Ec2CopyRouteTableAction

ec2_create_route_action

See Ec2CreateRouteAction

ec2_create_route_table_action

See Ec2CreateRouteTableAction

ec2_delete_route_action

See Ec2DeleteRouteAction

ec2_replace_route_action

See Ec2ReplaceRouteAction

ec2_replace_route_table_association_action

See Ec2ReplaceRouteTableAssociationAction

evaluation_result

See EvaluationResult

expected_route

See ExpectedRoute

network_firewall_black_hole_route_detected_violation

See NetworkFirewallBlackHoleRouteDetectedViolation

network_firewall_internet_traffic_not_inspected_violation

See NetworkFirewallInternetTrafficNotInspectedViolation

network_firewall_invalid_route_configuration_violation

See NetworkFirewallInvalidRouteConfigurationViolation

network_firewall_missing_expected_routes_violation

See NetworkFirewallMissingExpectedRoutesViolation

network_firewall_missing_expected_rt_violation

See NetworkFirewallMissingExpectedRtViolation

network_firewall_missing_firewall_violation

See NetworkFirewallMissingFirewallViolation

network_firewall_missing_subnet_violation

See NetworkFirewallMissingSubnetViolation

network_firewall_policy_description

See NetworkFirewallPolicyDescription

network_firewall_policy_modified_violation

See NetworkFirewallPolicyModifiedViolation

network_firewall_unexpected_firewall_routes_violation

See NetworkFirewallUnexpectedFirewallRoutesViolation

network_firewall_unexpected_gateway_routes_violation

See NetworkFirewallUnexpectedGatewayRoutesViolation

partial_match

See PartialMatch

policy

See Policy

policy_compliance_detail

See PolicyComplianceDetail

policy_compliance_status

See PolicyComplianceStatus

policy_summary

See PolicySummary

possible_remediation_action

See PossibleRemediationAction

possible_remediation_actions

See PossibleRemediationActions

protocols_list_data

See ProtocolsListData

protocols_list_data_summary

See ProtocolsListDataSummary

remediation_action

See RemediationAction

remediation_action_with_order

See RemediationActionWithOrder

resource_tag

See ResourceTag

resource_violation

See ResourceViolation

route

See Route

security_group_remediation_action

See SecurityGroupRemediationAction

security_group_rule_description

See SecurityGroupRuleDescription

security_service_policy_data

See SecurityServicePolicyData

stateful_rule_group

See StatefulRuleGroup

stateless_rule_group

See StatelessRuleGroup

tag

See Tag

violation_detail

See ViolationDetail

Structs

ActionTarget

Describes a remediation action target.

App

An individual Firewall Manager application.

AppsListData

An Firewall Manager applications list.

AppsListDataSummary

Details of the Firewall Manager applications list.

AwsEc2InstanceViolation

Violation detail for an EC2 instance resource.

AwsEc2NetworkInterfaceViolation

Violation detail for network interfaces associated with an EC2 instance.

AwsVpcSecurityGroupViolation

Violation detail for the rule violation in a security group when compared to the primary security group of the Firewall Manager policy.

ComplianceViolator

Details of the resource that is not protected by the policy.

DnsDuplicateRuleGroupViolation

A DNS Firewall rule group that Firewall Manager

DnsRuleGroupLimitExceededViolation

The VPC that Firewall Manager was applying a DNS Fireall policy to reached the limit for associated DNS Firewall rule groups. Firewall Manager tried to associate another rule group with the VPC and failed due to the limit.

DnsRuleGroupPriorityConflictViolation

A rule group that Firewall Manager

Ec2AssociateRouteTableAction

The action of associating an EC2 resource, such as a subnet or internet gateway, with a route table.

Ec2CopyRouteTableAction

An action that copies the EC2 route table for use in remediation.

Ec2CreateRouteAction

Information about the CreateRoute action in Amazon EC2.

Ec2CreateRouteTableAction

Information about the CreateRouteTable action in Amazon EC2.

Ec2DeleteRouteAction

Information about the DeleteRoute action in Amazon EC2.

Ec2ReplaceRouteAction

Information about the ReplaceRoute action in Amazon EC2.

Ec2ReplaceRouteTableAssociationAction

Information about the ReplaceRouteTableAssociation action in Amazon EC2.

EvaluationResult

Describes the compliance status for the account. An account is considered noncompliant if

ExpectedRoute

Information about the expected route in the route table.

NetworkFirewallBlackHoleRouteDetectedViolation

Violation detail for an internet gateway route with an inactive state in the customer subnet route table or Network Firewall subnet route table.

NetworkFirewallInternetTrafficNotInspectedViolation

Violation detail for the subnet for which internet traffic that hasn't been inspected.

NetworkFirewallInvalidRouteConfigurationViolation

Violation detail for the improperly configured subnet route. It's possible there is a missing route table route,

NetworkFirewallMissingExpectedRoutesViolation

Violation detail for an expected route missing in Network Firewall.

NetworkFirewallMissingExpectedRtViolation

Violation detail for Network Firewall for a subnet that's not associated to the expected

NetworkFirewallMissingFirewallViolation

Violation detail for Network Firewall for a subnet that doesn't have a

NetworkFirewallMissingSubnetViolation

Violation detail for Network Firewall for an Availability Zone that's

NetworkFirewallPolicyDescription

The definition of the Network Firewall firewall policy.

NetworkFirewallPolicyModifiedViolation

Violation detail for Network Firewall for a firewall policy that has a different

NetworkFirewallUnexpectedFirewallRoutesViolation

Violation detail for an unexpected route that's present in a route table.

NetworkFirewallUnexpectedGatewayRoutesViolation

Violation detail for an unexpected gateway route that’s present in a route table.

PartialMatch

The reference rule that partially matches the ViolationTarget rule and violation reason.

Policy

An Firewall Manager policy.

PolicyComplianceDetail

Describes

PolicyComplianceStatus

Indicates whether the account is compliant with the specified policy. An account is

PolicySummary

Details of the Firewall Manager policy.

PossibleRemediationAction

A list of remediation actions.

PossibleRemediationActions

A list of possible remediation action lists. Each individual possible remediation action is a list of individual remediation actions.

ProtocolsListData

An Firewall Manager protocols list.

ProtocolsListDataSummary

Details of the Firewall Manager protocols list.

RemediationAction

Information about an individual action you can take to remediate a violation.

RemediationActionWithOrder

An ordered list of actions you can take to remediate a violation.

ResourceTag

The resource tags that Firewall Manager uses to determine if a particular resource

ResourceViolation

Violation detail based on resource type.

Route

Describes a route in a route table.

SecurityGroupRemediationAction

Remediation option for the rule specified in the ViolationTarget.

SecurityGroupRuleDescription

Describes a set of permissions for a security group rule.

SecurityServicePolicyData

Details about the security service that is being used to protect the resources.

StatefulRuleGroup

Network Firewall stateful rule group, used in a NetworkFirewallPolicyDescription.

StatelessRuleGroup

Network Firewall stateless rule group, used in a NetworkFirewallPolicyDescription.

Tag

A collection of key:value pairs associated with an Amazon Web Services resource. The key:value pair can be anything you define. Typically, the tag key represents a category (such as "environment") and the tag value represents a specific value within that category (such as "test," "development," or "production"). You can add up to 50 tags to each Amazon Web Services resource.

ViolationDetail

Violations for a resource based on the specified Firewall Manager policy and Amazon Web Services account.

Enums

AccountRoleStatus

CustomerPolicyScopeIdType

DependentServiceName

DestinationType

PolicyComplianceStatusType

RemediationActionType

SecurityServiceType

TargetType

ViolationReason