Module aws_sdk_fms::operation

Structs

AssociateAdminAccount

Sets the Firewall Manager administrator account. The account must be

DeleteAppsList

Permanently deletes an Firewall Manager applications list.

DeleteNotificationChannel

Deletes an Firewall Manager association with the IAM role and the Amazon Simple

DeletePolicy

Permanently deletes an Firewall Manager policy.

DeleteProtocolsList

Permanently deletes an Firewall Manager protocols list.

DisassociateAdminAccount

Disassociates the account that has been set as the Firewall Manager administrator

GetAdminAccount

Returns the Organizations account that is associated with Firewall Manager

GetAppsList

Returns information about the specified Firewall Manager applications list.

GetComplianceDetail

Returns detailed compliance information about the specified member account. Details

GetNotificationChannel

Information

GetPolicy

Returns information about the specified Firewall Manager policy.

GetProtectionStatus

If you created a Shield Advanced policy, returns policy-level attack summary information

GetProtocolsList

Returns information about the specified Firewall Manager protocols list.

GetViolationDetails

Retrieves violations for a resource based on the specified Firewall Manager policy and Amazon Web Services account.

ListAppsLists

Returns an array of AppsListDataSummary objects.

ListComplianceStatus

Returns an array of PolicyComplianceStatus objects. Use

ListMemberAccounts

Returns a MemberAccounts object that lists the member accounts in the

ListPolicies

Returns an array of PolicySummary objects.

ListProtocolsLists

Returns an array of ProtocolsListDataSummary objects.

ListTagsForResource

Retrieves the list of tags for the specified Amazon Web Services resource.

PutAppsList

Creates an Firewall Manager applications list.

PutNotificationChannel

Designates the IAM role and Amazon Simple Notification Service (SNS) topic that

PutPolicy

Creates an Firewall Manager policy.

PutProtocolsList

Creates an Firewall Manager protocols list.

TagResource

Adds one or more tags to an Amazon Web Services resource.

UntagResource

Removes one or more tags from an Amazon Web Services resource.